Add function to save a TVDB token
The token is valid for 24-hours so no need to retrieve one each time the
client is authenticated.
Add function to use the saved token if it is still valid.
Add function to determine if a saved token is valid.
Add model and table to hold the TVDB token.

Rename TVDBClient to TvdbClient for consistency with TvdbToken.