Change how a show's tvdb info was set
Create update_tvdb action to update a show's tvdb info given a json
retrieved from the tvdb api.
This allows the controller to handle the contents of the json.
Change shows' search view's form to route to update_tvdb.

Remove forms to set the tvdb id manually since this should should be
handled by the search action.
Add a button to manually call the search action, if the current show
exists - since the search action requires a valid show to update.

Change the class used in the application layout's alert div to the
'alert-danger' class.