Add buttons to upgrade a user to an admin in the users index
Clicking a user button in the users index will change the associated
user to an admin.
Modify the user and delete buttons to appear disabled if the associated
user is an admin, since admins are not able to modify other admins.

Modify the update and destroy actions to not succeed if the requested
user is an admin and not the current user.

Add link to admin tools to the navbar if the signed in user is an
admin.
Add link to allow the signed in user to edit their account in the
navbar.