Add function to request admin access if user is not admin
If the user is not an admin, they can request admin access by clicking
the request admin link in to top right corner. This link would send an
email to the site owner regarding the request.

Add UserMailer to handle sending the request admin email. This does not
subclass devise's mailer since it doesn't really relate to devise
core functionality.