Add api namespace and controller to serve requests through the STW api
Update the routes to include api specific routes.
Move the hardcoded shows routes to the shows resource block.
Update the navbar with the new routes' paths.

Add api controller to ignore csrf token, since the check is likely
not needed for json requests.
Add index, abandoned, completed, and in progress actions to the shows
api controller.