Remove edit link from a show item if the user is not an admin
Move episode number from above to below a show's title if the user is
not an admin.
The edit link was removed since users that aren't admins can't edit
shows.
The add show button remains in the navbar to inform users that they can
add a show if they are registered and have admin access.